Gimmel-bot/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-absence-bot"
version = "1.0.0"
description = "Discord bot for daily absence day reports"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Alexey Sarafanov" }]
dependencies = [
  "discord.py>=2.4.0,<3.0.0",
  "aiosqlite>=0.20.0,<1.0.0",
  "APScheduler>=3.10.4,<4.0.0",
  "aiohttp>=3.9.5,<4.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.23.8,<1.0.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
